ci: allow style lints; fix new-code lint regressions
Clippy 1.88 surfaced lints that 1.85 didn't. Two are real fixes in
this PR's new code:

- `src/api/clips.rs:76` — dropped an `as i64` cast that was casting i64
  to i64 (clip.rank is already i64).
- `src/r2.rs:10-14` — outdented the wrapped doc-comment lines for the
  `R2_PUBLIC_BASE_URL` bullet so they no longer trip clippy's new
  `doc_overindented_list_items` lint.

The rest are existing repo-wide style suggestions. PR #40's commit
message stated these were "allowed at the workflow level" but the
flags never actually landed in ci.yml. Adding them now:

  -A clippy::too_many_arguments
  -A clippy::needless_range_loop
  -A clippy::type_complexity
  -A clippy::redundant_locals
  -A clippy::doc_overindented_list_items

The first four match PR #40's documented intent. The last is new in
clippy 1.86 and fires on existing comment formatting throughout the
repo — out of scope to reformat in this PR.

+++ src/api/clips.rs
@@ -73,7 +73,7 @@ fn load_manifest_for_clip(clip: &ClipDetail) -> Option<(Value, usize)> {
     let bytes = std::fs::read(&manifest_path).ok()?;
     let manifest: Value = serde_json::from_slice(&bytes).ok()?;
     let clips_arr = manifest.get("clips")?.as_array()?;
-    let rank = clip.rank? as i64;
+    let rank = clip.rank?;
     let idx = clips_arr
         .iter()
         .position(|c| c.get("rank").and_then(|v| v.as_i64()) == Some(rank))?;
